feat: check nm-applet before kill

+++ bin/disable-nmapplet.sh
@@ -1,8 +1,8 @@
 #!/usr/bin/env bash
 set -e
 
-sudo sh -c 'echo "Hidden=true" >> /etc/xdg/autostart/nm-applet.desktop'
-
-killall -9 nm-applet
-
-sudo pacman -Rns nm-applet
+if pacman -Qi networkmanager-applet >/dev/null; then
+    sudo sh -c 'echo "Hidden=true" >> /etc/xdg/autostart/nm-applet.desktop'
+    killall -9 nm-applet
+    sudo pacman -Rns --noconfirm networkmanager-applet
+fi
